Create Disabled_User_Licenses.ps1
Gets a list of users that are in specific groups and disabled and outputs them to a csv file. The while loop can be much more optimized by having the foreach loop check for the SamAccountName and output that list to the arrays, as opposed to the while loop checking for the SamAccountName.